chore: move tests for arithmetic generics closer to the code
