+++ compiler/noirc_frontend/proptest-regressions/hir_def/types/arithmetic.txt
