+++ cspell.json
@@ -39,6 +39,7 @@
         "callstack",
         "callstacks",
         "canonicalize",
+        "canonicalized",
         "castable",
         "catmcgee",
         "Celo",
@@ -173,6 +174,7 @@
         "prettytable",
         "printstd",
         "proptest",
+        "proptests",
         "pseudocode",
         "pubkey",
         "quantile",
@@ -199,8 +201,8 @@
         "stdlib",
         "structs",
         "subexpression",
-        "subtrait",
         "subshell",
+        "subtrait",
         "subtyping",
         "supertrait",
         "supertraits",
